+++ res/shaders/face.vert
@@ -1,10 +1,10 @@
 #version 450 core
 layout (location = 0) in vec3 aPos;
 layout (location = 1) in vec2 aUV;
-layout (location = 2) in int aTexIndex;
+layout (location = 2) in float aTexIndex;
 
 out vec2 uv;
-out flat int texIndex; 
+out flat float texIndex; 
 
 uniform mat4 aVP;
